fix: To use complete range for catalog entries index internal used datatype was switched from short to int.

+++ src/DataService/Rest/CatalogEntry.cs
@@ -59,7 +59,7 @@ namespace Zeiss.IMT.PiWeb.Api.DataService.Rest
 		/// Gets or sets the unique key of this catalog entry.
 		/// </summary>
 		[JsonProperty( "key" )]
-		public short Key { get; set; }
+		public int Key { get; set; }
 
 		#endregion
